requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bleach==6.0.0
cached-property==1.5.2
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.2.0
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.103.2
fastjsonschema==2.16.2
fileupload==0.1.5
Flask==2.0.3
fonttools==4.38.0
frozenlist==1.3.3
git-filter-repo==2.45.0
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-server==1.23.6
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.5
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
null==0.6.1
numpy==1.21.6
openai==0.28.0
openpyxl==3.1.3
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psutil==5.9.4
psycopg==3.1.20
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.9.0
pydantic_core==2.14.6
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.12
PyQt5-Qt5==5.15.2
PyQt5-stubs==5.15.6.0
PyQt5_sip==4.19.19
pyrsistent==0.19.3
python-chess==1.999
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2024.2
pyzmq==25.0.0
requests==2.31.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.5
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.21.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
whitenoise==6.5.0
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
yarl==1.9.4
zipp==3.14.0
